Innført ElasticsearchRepository fra Spring Data. Bruker det ikke til noe særlig ennå, men kan være kjekt å ha
