+++ src/main/kotlin/no/nav/klage/oppgave/config/ElasticsearchServiceConfiguration.kt
@@ -1,20 +1,28 @@
 package no.nav.klage.oppgave.config
 
-import no.nav.klage.oppgave.repositories.ElasticsearchRepository
+import no.nav.klage.oppgave.repositories.EsKlagebehandlingRepository
 import no.nav.klage.oppgave.repositories.InnloggetSaksbehandlerRepository
+import no.nav.klage.oppgave.service.ElasticsearchService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.elasticsearch.core.ElasticsearchRestTemplate
+import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories
 
 @Configuration
+@EnableElasticsearchRepositories(basePackageClasses = [EsKlagebehandlingRepository::class])
 class ElasticsearchServiceConfiguration {
 
     @Bean
     fun elasticsearchRepository(
         elasticsearchRestTemplate: ElasticsearchRestTemplate,
-        innloggetSaksbehandlerRepository: InnloggetSaksbehandlerRepository
-    ): ElasticsearchRepository {
-        return ElasticsearchRepository(elasticsearchRestTemplate, innloggetSaksbehandlerRepository)
+        innloggetSaksbehandlerRepository: InnloggetSaksbehandlerRepository,
+        esKlagebehandlingRepository: EsKlagebehandlingRepository
+    ): ElasticsearchService {
+        return ElasticsearchService(
+            elasticsearchRestTemplate,
+            innloggetSaksbehandlerRepository,
+            esKlagebehandlingRepository
+        )
     }
 
 }
