+++ src/main/kotlin/no/nav/klage/oppgave/repositories/EsKlagebehandlingRepository.kt
@@ -1,0 +1,7 @@
+package no.nav.klage.oppgave.repositories
+
+import no.nav.klage.oppgave.domain.elasticsearch.EsKlagebehandling
+import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
+
+interface EsKlagebehandlingRepository : ElasticsearchRepository<EsKlagebehandling, String> {
+}
