+++ src/test/kotlin/no/nav/klage/oppgave/service/StatistikkInElasticsearchServiceTest.kt
@@ -2,11 +2,9 @@ package no.nav.klage.oppgave.service
 
 import com.ninjasquad.springmockk.MockkBean
 import no.nav.klage.oppgave.config.ElasticsearchServiceConfiguration
-import no.nav.klage.oppgave.domain.KlagebehandlingerSearchCriteria
 import no.nav.klage.oppgave.domain.elasticsearch.EsKlagebehandling
 import no.nav.klage.oppgave.domain.kodeverk.Sakstype
 import no.nav.klage.oppgave.domain.kodeverk.Tema
-import no.nav.klage.oppgave.repositories.ElasticsearchRepository
 import no.nav.klage.oppgave.repositories.InnloggetSaksbehandlerRepository
 import org.assertj.core.api.Assertions.assertThat
 import org.elasticsearch.client.RestHighLevelClient
@@ -18,6 +16,7 @@ import org.junit.jupiter.api.TestMethodOrder
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration
 import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchDataAutoConfiguration
+import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchRepositoriesAutoConfiguration
 import org.springframework.boot.autoconfigure.elasticsearch.ElasticsearchRestClientAutoConfiguration
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.data.elasticsearch.core.ElasticsearchRestTemplate
@@ -39,8 +38,12 @@ import java.time.LocalDate
 @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
 @Testcontainers
 @SpringBootTest(classes = [ElasticsearchServiceConfiguration::class])
-@ImportAutoConfiguration(ElasticsearchRestClientAutoConfiguration::class, ElasticsearchDataAutoConfiguration::class)
-class ElasticsearchRepositoryTest {
+@ImportAutoConfiguration(
+    ElasticsearchRestClientAutoConfiguration::class,
+    ElasticsearchDataAutoConfiguration::class,
+    ElasticsearchRepositoriesAutoConfiguration::class
+)
+class StatistikkInElasticsearchServiceTest {
 
     companion object {
         @Container
@@ -64,7 +67,7 @@ class ElasticsearchRepositoryTest {
     lateinit var innloggetSaksbehandlerRepository: InnloggetSaksbehandlerRepository
 
     @Autowired
-    lateinit var repository: ElasticsearchRepository
+    lateinit var service: ElasticsearchService
 
     @Autowired
     lateinit var esTemplate: ElasticsearchRestTemplate
@@ -86,40 +89,47 @@ class ElasticsearchRepositoryTest {
         assertThat(indexOps.exists()).isTrue()
     }
 
-    @Test
-    @Order(3)
-    fun `lagrer to oppgaver for senere tester`() {
-
-
-        val klagebehandling1 = EsKlagebehandling(
-            id = "1001L",
+    private fun klagebehandling(id: Long, innsendt: LocalDate, frist: LocalDate, avsluttet: LocalDate? = null) =
+        EsKlagebehandling(
+            id = id.toString(),
             versjon = 1L,
             tildeltEnhet = "4219",
-            tema = Tema.SYK,
+            tema = Tema.FOR,
             sakstype = Sakstype.KLAGE,
             tildeltSaksbehandlerident = null,
-            innsendt = LocalDate.of(2019, 10, 1),
-            mottattFoersteinstans = LocalDate.of(2019, 11, 1),
-            mottattKlageinstans = LocalDate.of(2019, 12, 1),
-            frist = LocalDate.of(2020, 12, 1),
+            innsendt = innsendt,
+            mottattFoersteinstans = LocalDate.of(2018, 11, 1),
+            mottattKlageinstans = LocalDate.of(2018, 12, 1),
+            frist = frist,
+            avsluttet = avsluttet,
             hjemler = listOf()
         )
-        val klagebehandling2 =
-            EsKlagebehandling(
-                id = "1002L",
-                versjon = 1L,
-                tildeltEnhet = "4219",
-                tema = Tema.FOR,
-                sakstype = Sakstype.KLAGE,
-                tildeltSaksbehandlerident = null,
-                innsendt = LocalDate.of(2018, 10, 1),
-                mottattFoersteinstans = LocalDate.of(2018, 11, 1),
-                mottattKlageinstans = LocalDate.of(2018, 12, 1),
-                frist = LocalDate.of(2019, 12, 1),
-                hjemler = listOf()
-            )
-        esTemplate.save(klagebehandling1)
-        esTemplate.save(klagebehandling2)
+
+    @Test
+    @Order(3)
+    fun `saving klagebehandlinger for later tests`() {
+
+
+        fun idag() = LocalDate.now()
+        fun uviktigdag() = LocalDate.now()
+
+        val klagebehandlinger = listOf(
+            klagebehandling(1001L, idag(), idag()),
+            klagebehandling(1002L, idag().minusDays(1), idag().minusDays(1)),
+            klagebehandling(1003L, idag().minusDays(6), idag().plusDays(1)),
+            klagebehandling(1004L, idag().minusDays(7), idag().minusDays(7)),
+            klagebehandling(1005L, idag().minusDays(8), idag().plusDays(7)),
+            klagebehandling(1006L, idag().minusDays(30), idag().minusDays(30)),
+            klagebehandling(1007L, idag().minusDays(31), idag().plusDays(30)),
+            klagebehandling(2001L, idag(), idag(), idag()),
+            klagebehandling(2002L, idag().minusDays(1), uviktigdag(), idag().minusDays(1)),
+            klagebehandling(2003L, idag().minusDays(6), uviktigdag(), idag().minusDays(6)),
+            klagebehandling(2004L, idag().minusDays(7), uviktigdag(), idag().minusDays(7)),
+            klagebehandling(2005L, idag().minusDays(8), uviktigdag(), idag().minusDays(8)),
+            klagebehandling(2006L, idag().minusDays(30), uviktigdag(), idag().minusDays(30)),
+            klagebehandling(2007L, idag().minusDays(31), uviktigdag(), idag().minusDays(31)),
+        )
+        esTemplate.save(klagebehandlinger)
 
         sleep(2000L)
 
@@ -127,37 +137,22 @@ class ElasticsearchRepositoryTest {
             .withQuery(QueryBuilders.matchAllQuery())
             .build()
         val searchHits: SearchHits<EsKlagebehandling> = esTemplate.search(query, EsKlagebehandling::class.java)
-        assertThat(searchHits.totalHits).isEqualTo(2L)
+        assertThat(searchHits.totalHits).isEqualTo(14L)
     }
 
     @Test
     @Order(4)
-    fun `Klagebehandling can be searched for by tema`() {
-        val klagebehandlinger: List<EsKlagebehandling> =
-            repository.findByCriteria(
-                KlagebehandlingerSearchCriteria(
-                    temaer = listOf(Tema.SYK),
-                    offset = 0,
-                    limit = 10
-                )
-            ).searchHits.map { it.content }
-        assertThat(klagebehandlinger.size).isEqualTo(1L)
-        assertThat(klagebehandlinger.first().id).isEqualTo("1001L")
+    fun `Klagebehandling statistikk gives correct numbers`() {
+        val statistikkTall = service.statistikkQuery()
+        assertThat(statistikkTall.ubehandlede).isEqualTo(7)
+        assertThat(statistikkTall.overFrist).isEqualTo(3)
+        assertThat(statistikkTall.innsendtIGaar).isEqualTo(2)
+        assertThat(statistikkTall.innsendtSiste7Dager).isEqualTo(6)
+        assertThat(statistikkTall.innsendtSiste30Dager).isEqualTo(10)
+        assertThat(statistikkTall.behandletIGaar).isEqualTo(1)
+        assertThat(statistikkTall.behandletSiste7Dager).isEqualTo(3)
+        assertThat(statistikkTall.behandletSiste30Dager).isEqualTo(5)
     }
 
-    @Test
-    @Order(5)
-    fun `Klagebehandling can be searched for by frist`() {
-        val klagebehandlinger: List<EsKlagebehandling> =
-            repository.findByCriteria(
-                KlagebehandlingerSearchCriteria(
-                    fristFom = LocalDate.of(2020, 12, 1),
-                    offset = 0,
-                    limit = 10
-                )
-            ).searchHits.map { it.content }
-        assertThat(klagebehandlinger.size).isEqualTo(1L)
-        assertThat(klagebehandlinger.first().id).isEqualTo("1001L")
-    }
 
 }
